Day 8 Temperature Conversion
On the day 8 of 100 days learning challenge.

I have created a temperature Converter (Inspired by Brocode)and here we have an index.html file with a form tag and it contains radio buttons for selecting either Celsius or Fahrenheit.

Then we take those inputs from the HTML file and access them in the script.js and create a function which executes the logic part for converting the temperatures using If Else statements.

Finally, styled it using CSS.